+++ frontend/src/components/TicketRightDrawerItems/EditButton.tsx
@@ -1,4 +1,4 @@
-import Typography from "@mui/material/Typography";
+import Button from "@mui/material/Button";
 
 interface EditButtonProps {
   onStartEdit: any;
@@ -8,14 +8,9 @@ const EditButton = (props: EditButtonProps): JSX.Element => {
   const { onStartEdit } = props;
 
   return (
-    <Typography
-      variant='body2'
-      color='primary'
-      sx={{ "&:hover": { textDecoration: "underline", cursor: "pointer" } }}
-      onClick={onStartEdit}
-    >
+    <Button variant='text' size='small' color='primary' onClick={onStartEdit}>
       Edit
-    </Typography>
+    </Button>
   );
 };