+++ frontend/src/components/TicketRightDrawerItems/PriorityEditItem.tsx
@@ -2,7 +2,6 @@ import Box from "@mui/material/Box";
 import Chip from "@mui/material/Chip";
 import Divider from "@mui/material/Divider";
 import ListItem from "@mui/material/ListItem";
-import Typography from "@mui/material/Typography";
 import { useState } from "react";
 
 import useProductBacklog from "../../hooks/useProductBacklog";
@@ -11,6 +10,7 @@ import { capitalizeFirstLetter } from "../../utils/capitalizeFirstLetter";
 import FormPriorityToggleButtons from "../FormPriorityToggleButtons";
 import PriorityIcon from "../PriorityIcon";
 import EditButton from "./EditButton";
+import RightDrawerTitle from "./RightDrawerTitle";
 import UpdateButton from "./UpdateButton";
 
 interface PriorityEditItemProps {
@@ -38,10 +38,12 @@ const PriorityEditItem = (props: PriorityEditItemProps): JSX.Element => {
   if (isEditModeOn) {
     return (
       <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-        <Box display='flex' width='100%' mb={1} gap={1}>
-          <Typography variant='body2'>Priority</Typography>
-          <UpdateButton onAccept={handleChangePriority} onCancel={handleToggleEditMode} showUpdateButton={false} />
-        </Box>
+        <RightDrawerTitle
+          title='Priority'
+          actionButton={
+            <UpdateButton onAccept={handleChangePriority} onCancel={handleToggleEditMode} showSaveButton={false} />
+          }
+        />
         <Box mb={2}>
           <FormPriorityToggleButtons value={priority} onChangeHandler={handleChangePriority} />
         </Box>
@@ -52,10 +54,7 @@ const PriorityEditItem = (props: PriorityEditItemProps): JSX.Element => {
 
   return (
     <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-      <Box display='flex' gap={1} width='100%' mb={1}>
-        <Typography variant='body2'>Priority</Typography>
-        <EditButton onStartEdit={handleToggleEditMode} />
-      </Box>
+      <RightDrawerTitle title='Priority' actionButton={<EditButton onStartEdit={handleToggleEditMode} />} />
       <Chip
         label={`${capitalizeFirstLetter(priority)} Priority`}
         icon={