+++ frontend/src/components/TicketRightDrawerItems/SubtaskEditItem.tsx
@@ -1,4 +1,4 @@
-import { Box, Divider, ListItem, ListItemText, TextField, Typography } from "@mui/material";
+import { Box, Divider, ListItem, ListItemText, TextField } from "@mui/material";
 import { useEffect, useState } from "react";
 
 import fetchSubtasksByIds from "../../api/tickets/fetchSubtasksByIds";
@@ -7,8 +7,9 @@ import useProductBacklog from "../../hooks/useProductBacklog";
 import Subtask from "../../types/Subtask";
 import TicketSubtask from "../TicketSubtask";
 import TicketSubtaskDeletable from "../TicketSubtaskDeletable";
-import TextButton from "../common/TextButton";
 import EditButton from "./EditButton";
+import RightDrawerTitle from "./RightDrawerTitle";
+import UpdateButton from "./UpdateButton";
 
 interface SubtaskEditItemProps {
   subtaskIds: string[];
@@ -97,17 +98,18 @@ const SubtaskEditItem = (props: SubtaskEditItemProps): JSX.Element => {
   if (isEditModeOn) {
     return (
       <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-        <Box display='flex' gap={1} width='100%' mb={1}>
-          <Typography variant='body2'>Subtasks</Typography>
-          <TextButton handleOnClick={handleToggleEditMode} text='Done' />
-        </Box>
+        <RightDrawerTitle
+          title='Subtasks'
+          actionButton={
+            <UpdateButton onAccept={handleToggleEditMode} onCancel={handleToggleEditMode} showSaveButton={false} />
+          }
+        />
         <form onSubmit={(e) => handleAddSubtask(e)} style={{ width: "100%" }}>
           <TextField
-            size='small'
             variant='filled'
             autoFocus
             fullWidth
-            inputProps={{ style: { padding: 7, fontSize: 14 } }}
+            inputProps={{ style: { padding: 7 } }}
             placeholder='Type and press enter to add subtask'
             value={titleInput}
             onChange={(e) => setTitleInput(e.target.value)}
@@ -130,10 +132,7 @@ const SubtaskEditItem = (props: SubtaskEditItemProps): JSX.Element => {
 
   return (
     <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-      <Box display='flex' gap={1} width='100%'>
-        <Typography variant='body2'>Subtasks</Typography>
-        <EditButton onStartEdit={handleToggleEditMode} />
-      </Box>
+      <RightDrawerTitle title='Subtasks' actionButton={<EditButton onStartEdit={handleToggleEditMode} />} />
       {subtaskIds.length === 0 && <ListItemText secondary='None' sx={{ mt: 0.5 }} />}
       {subtasks.length > 0 && (
         <Box maxWidth='100%' sx={{ mt: 1 }}>