+++ frontend/src/components/TicketRightDrawerItems/TicketTypeEditItem.tsx
@@ -2,7 +2,6 @@ import Box from "@mui/material/Box";
 import Chip from "@mui/material/Chip";
 import Divider from "@mui/material/Divider";
 import ListItem from "@mui/material/ListItem";
-import Typography from "@mui/material/Typography";
 import { useState } from "react";
 
 import useProductBacklog from "../../hooks/useProductBacklog";
@@ -11,6 +10,7 @@ import { capitalizeFirstLetter } from "../../utils/capitalizeFirstLetter";
 import FormTicketTypeToggleButtons from "../FormTicketTypeToggleButtons";
 import TicketTypeIcon from "../TicketTypeIcon";
 import EditButton from "./EditButton";
+import RightDrawerTitle from "./RightDrawerTitle";
 import UpdateButton from "./UpdateButton";
 
 interface TicketTypeEditItemProps {
@@ -38,10 +38,12 @@ const TicketTypeEditItem = (props: TicketTypeEditItemProps): JSX.Element => {
   if (isEditModeOn) {
     return (
       <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-        <Box display='flex' width='100%' mb={1} gap={1}>
-          <Typography variant='body2'>Type</Typography>
-          <UpdateButton onAccept={handleChangeTicketType} onCancel={handleToggleEditMode} showUpdateButton={false} />
-        </Box>
+        <RightDrawerTitle
+          title='Type'
+          actionButton={
+            <UpdateButton onAccept={handleChangeTicketType} onCancel={handleToggleEditMode} showSaveButton={false} />
+          }
+        />
         <Box mb={2}>
           <FormTicketTypeToggleButtons value={type} onChangeHandler={handleChangeTicketType} />
         </Box>
@@ -52,10 +54,7 @@ const TicketTypeEditItem = (props: TicketTypeEditItemProps): JSX.Element => {
 
   return (
     <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-      <Box display='flex' gap={1} width='100%' mb={1}>
-        <Typography variant='body2'>Type</Typography>
-        <EditButton onStartEdit={handleToggleEditMode} />
-      </Box>
+      <RightDrawerTitle title='Type' actionButton={<EditButton onStartEdit={handleToggleEditMode} />} />
       <Chip
         label={`${capitalizeFirstLetter(type)}`}
         icon={