+++ frontend/src/components/common/UserAvatar.tsx
@@ -1,6 +1,7 @@
 import { Avatar } from "@mui/material";
 
 import getUserAvatarSVG from "../../utils/getUserAvatarSVG";
+import Tooltip from "./Tooltip";
 
 interface UserAvatarProps {
   username: string;
@@ -11,9 +12,11 @@ const UserAvatar = (props: UserAvatarProps): JSX.Element => {
   const { username, displayName } = props;
 
   return (
-    <Avatar src={getUserAvatarSVG(username)} sx={{ height: 32, width: 32, backgroundColor: "grey.300" }}>
-      {displayName}
-    </Avatar>
+    <Tooltip title={displayName}>
+      <Avatar src={getUserAvatarSVG(username)} sx={{ height: 32, width: 32, backgroundColor: "grey.300" }}>
+        {displayName}
+      </Avatar>
+    </Tooltip>
   );
 };